+++ src/main/java/com/gym/gym/service/LockerService.java
@@ -1,7 +1,8 @@
 package com.gym.gym.service;
 
 import com.gym.gym.dto.LockerDto;
-import com.gym.gym.entity.Locker;
+import com.gym.gym.exception.exceptions.EntityNotFoundException;
+import com.gym.gym.exception.messages.ExceptionMessages;
 import com.gym.gym.mapper.LockerMapper;
 import com.gym.gym.repository.LockerRepository;
 import lombok.AllArgsConstructor;
@@ -15,19 +16,11 @@ import java.util.stream.Collectors;
 @Transactional
 @Service
 public class LockerService {
-    private final LockerRepository lockerRepository;
+    private LockerRepository lockerRepository;
 
     public List<LockerDto> getAllLockers() {
         return lockerRepository.findAll()
                 .stream()
-                .map(LockerMapper::toDto)
-                .collect(Collectors.toList());
-    }
-
-    public List<LockerDto> getAllAvailableLockers() {
-        return lockerRepository.findAll()
-                .stream()
-                .filter(Locker::isAvailable)
                 .map(LockerMapper::toDto)
                 .collect(Collectors.toList());
     }
@@ -37,6 +30,10 @@ public class LockerService {
     }
 
     public void delete(Long id) {
+
+        if (lockerRepository.findById(id).isEmpty()){
+            throw new EntityNotFoundException("Locker", id);
+        }
         lockerRepository.deleteById(id);
     }
